Extract hotspots state into custom hook
REDMINE-20853
